+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
 - Add plone.app.caching >= 3.0.0a1 dependency to make it work in Plone 5.2
   [erral]
 
+- Enable versioning.
+  [cekk]
 
 1.4.0 (2021-10-10)
 ------------------